+++ src/test/required.test.ts
@@ -1,6 +1,8 @@
-import { Position, Range } from "vscode-languageserver-types";
+import { Position, Range, DiagnosticSeverity } from "vscode-languageserver-types";
 import { createDiagnostic } from "../util";
 import { Test } from "./test";
+import { Validator } from "../validator";
+import { deepStrictEqual } from "assert";
 
 suite("Required settings for sections tests", () => {
     const tests: Test[] = [
@@ -323,4 +325,42 @@ suite("Required: No duplicate errors with [tags]", () => {
         Range.create(0, 1, 0, "widget]".length),
         "type is required",
     )]).validationTest();
+});
+
+suite("UDF settings tests", () => {
+  test("Required evaluate-expression is missing", () => {
+    const config = `[configuration]
+          [group]
+          [widget]
+              type = chart
+              metric = a
+              [column]
+                  [series]
+                  entity = b
+                  metrics = c, d, e`;
+    const validator = new Validator(config);
+    const actualDiagnostics = validator.lineByLine();
+    const expectedDiagnostic = createDiagnostic(
+      Range.create(Position.create(6, 19), Position.create(6, 25)),
+      "evaluate-expression is required if metrics is specified",
+      DiagnosticSeverity.Error
+    );
+    deepStrictEqual(actualDiagnostics, [expectedDiagnostic]);
+  });
+
+  test("No error: evaluate-expression is present", () => {
+    const config = `[configuration]
+          [group]
+          [widget]
+              type = chart
+              metric = a
+              [column]
+                  [series]
+                  entity = b
+                  evaluate-expression = c LIKE 'cpu*'
+                  metrics = c, d, e`;
+    const validator = new Validator(config);
+    const actualDiagnostics = validator.lineByLine();
+    deepStrictEqual(actualDiagnostics, []);
+  });
 });
