--- types/test/alias.test.d.ts
@@ -1,1 +1,0 @@
-export {};
