--- types/test/completionProvider.test.d.ts
@@ -1,1 +1,0 @@
-export {};
