--- types/test/deprecatedSettings.test.d.ts
@@ -1,1 +1,0 @@
-export {};
