--- types/test/endList.test.d.ts
@@ -1,1 +1,0 @@
-export {};
