--- types/test/forIn.test.d.ts
@@ -1,1 +1,0 @@
-export {};
