--- types/test/freemarker.test.d.ts
@@ -1,1 +1,0 @@
-export {};
