--- types/test/helpers/initResources.d.ts
@@ -1,1 +1,0 @@
-export {};
