--- types/test/helpers/resourcesProvider.d.ts
@@ -1,20 +1,0 @@
-import { Setting } from "../../setting";
-import { ResourcesProviderBase } from "../../resourcesProviderBase";
-export declare class ResourcesProvider extends ResourcesProviderBase {
-    constructor();
-    /**
-     * Reads snippets from "snippets.json" file
-     * @returns snippets JSON contents
-     */
-    readSnippets(): string;
-    /**
-     * Reads dictionary from "dictionary.json" file
-     * @returns array of settings from the file
-     */
-    protected readSettings(): Setting[];
-    /**
-     * Reads descriptions from "descriptions.md" file
-     * @returns map of settings names and descriptions
-     */
-    protected readDescriptions(): Map<string, string>;
-}
