--- types/test/ifCondition.test.d.ts
@@ -1,1 +1,0 @@
-export {};
