--- types/test/relatedSettings.test.d.ts
@@ -1,1 +1,0 @@
-export {};
