--- types/test/repetition.test.d.ts
@@ -1,1 +1,0 @@
-export {};
