--- types/test/scriptFormatting.test.d.ts
@@ -1,1 +1,0 @@
-export {};
