--- types/test/setting.test.d.ts
@@ -1,1 +1,0 @@
-export {};
