--- types/test/startEndRelation.test.d.ts
@@ -1,1 +1,0 @@
-export {};
