--- types/test/time/parseDate.test.d.ts
@@ -1,1 +1,0 @@
-export {};
