--- types/test/udfSettings.test.d.ts
@@ -1,1 +1,0 @@
-export {};
