--- types/test/undefinedVar.test.d.ts
@@ -1,1 +1,0 @@
-export {};
