--- types/test/var.test.d.ts
@@ -1,1 +1,0 @@
-export {};
